fetch/metadata/generated/js-module-via-dynamic-import-sec-fetch-mode-module-redirect.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.js"></script>
  <script type="module">
  'use strict';

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return import('https://{{hosts[alt][]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' + path + '&mime=application/javascript')
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
    });
  }, 'sec-fetch-mode - Cross-Site -> Same-Origin');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return import('https://{{hosts[alt][]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{domains[www]}}:{{ports[https][0]}}' + path + '&mime=application/javascript')
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
    });
  }, 'sec-fetch-mode - Cross-Site -> Same-Site');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return import('https://{{hosts[alt][]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{hosts[alt][]}}:{{ports[https][0]}}' + path + '&mime=application/javascript')
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
    });
  }, 'sec-fetch-mode - Cross-Site -> Cross-Site');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return import('https://{{host}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{hosts[alt][]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' + path + '&mime=application/javascript')
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
    });
  }, 'sec-fetch-mode - Same-Origin -> Cross-Site -> Same-Origin');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return import('https://{{host}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=http://{{host}}:{{ports[http][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' + path + '&mime=application/javascript')
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
    });
  }, 'sec-fetch-mode - HTTPS downgrade-upgrade');
  </script>
</html>
